PALLADIO-432
The StoEx analyzer (de.uka.ipd.sdq.stoex.analyser) is used to do several analyses on stoex such as type inference. These features would be better located at the commons project, where it can be used by other users of stoex as well.

The only reason for the bundle to be located at the PCM repo is ExpressionInferTypeVisitor::caseVariable. The method contains some special handling of variable characterizations. This is, however, only a special case of the variable handling. It would be better to make ExpressionInferTypeVisitor extensible and provide an extension from the PCM bundle.